feat: Request validator now includes email address validation
The existing validation allows malformed email addresses to pass validation.
This change adds a simple validation mechanism via the inbuilt DataAnnotations
which is simple and does not rely on exceptions like methods in System.Net.Mail.
Note that this solution would consider null as a valid email address, however,
this is covered with the required field validation.