tests: Clear separation between doctor email tests
Removed null and empty test cases from doctor request validator tests. Email
address validation is a different test to checking of the existence of the
field. Doing this allows a more specific error message to be provided on error.
Null or empty checking is already covered in a dedicated test.