tests: Improve coverage of doctor availability test
Added parameterized test cases to test doctor time slot availability. There are
4 simple scenarios we minimally need to account for, a perfect double booking,
a time slot requested that falls fully inside an already booked appointment,
a request that overlaps with the start of an existing booking and finally one
that overlaps with the end. For the sake of these tests I have assumed that
times that are equal are in the same slot.

The tests here are probably missing lots of edge cases, for example, what if
the doctor had multiple bookings inside a patient requested time slot?

Finally, setting the static dates in the doctor booking test way into the future
feels like an awful hack to get around the issues with the use of DateTime.Now
in the new booking validation code. This feels like booking time validation and
doctor availability should be refactored to be separate.