feat: Adds doctor timeslot checking to new booking validation
This uses a very naive check of timeslot intersection. We already know from the
other validation added earlier that a start time is guarenteed to be before an
end time, so we can use a simple method to check for overlap. I have assumed
that a booking can be of any length.  As previously put in the unit test commit,
there are likely edge cases here that have not been met that would need more
investigation if more time were to be taken.