feat: Adds boilerplate for a BookingService
In preparation for adding a cancellation endpoint a new BookingService has been
stood up and injected into the BookingController. The BookingService will
contain the business logic for the cancellation request, leaving the controller
endpoint purely to handle the HTTP request. This decoupling and injection will
allow the controller and BookingService to be unit tested.

All the other business logic in the BookingController should be refactored to
use the same Booking service.