feat: Moves MyOrderResult into dedicated file
Moves MyOrderResult into a dedicated file and placed into a "Responses" folder
under "BookingServices". With the introduction of BookingServices the Booking
Controller can now be cleaned up to only contain code relevant to the HTTP
interactions with the user. The location chosen for the new file mirrors the
locations used for the clinic/doctor/patient services, which aids project
familiarity.