feat: Populates BookingService with functionality
This duplicates the business logic contained in the BookingController into the
BookingService and alters it where appropriate to return true/false results
on success/failure. The controller still needs to know if these operations have
been successful or not for HTTP responses. If multiple states could be returned
from these operations an alternative way of reporting errors would need to be
used.

In additional a small POCO has been created to model the response to getting a
patients next appointment.