feat: Refactors BookingController to user BookingService
Removes all existing business logic from the BookingController and replaces with
the functionlity migrated into the BookingService. This leaves the controller
looking much cleaner.

Note that UpdateLatestBooking is still in the controller class. The purpose of
this method is not quite clear yet and may come into play when implementing
the cancellation feature at a later time.