tests: Adds tests got the BookingService
Basic tests to cover the main use cases for the cancellation feature. It seems
that there are 2 ways to approach the cancellation, either delete bookings
entirely or add a new property to capture the state of a booking. I have decided
to go with the latter and never delete any data, the ticket doesn't outright
state that this is required, but updates to GetPatientNextAppointment wouldn't
be required if we just deleted the bookings from the database.