feat: Adds functionality to cancel bookings
Adds a fleshed out CancelBooking method to the BookingService. Simple checks are
performed to checking if a booking exists and that the patient and booking
match one another. A cancelled booking has a property set to signify that the
booking is no longer required.

GetPatientNextAppointment includes a simple filter operation to remove any
results that have the cancelled property set to true.